+++ wand/cdefs/structures.py
@@ -7,8 +7,8 @@ from ctypes import POINTER, Structure, c_bool, c_double, c_int, c_size_t
 from wand.cdefs.wandtypes import c_ssize_t, c_magick_real_t, c_magick_size_t
 
 __all__ = ('AffineMatrix', 'CCMaxMetrics', 'CCObjectInfo', 'CCObjectInfo70A',
-           'ChannelFeature', 'GeometryInfo', 'KernelInfo', 'MagickPixelPacket',
-           'PixelInfo', 'PointInfo', 'RectangleInfo')
+           'CCObjectInfo711', 'ChannelFeature', 'GeometryInfo', 'KernelInfo',
+           'MagickPixelPacket', 'PixelInfo', 'PointInfo', 'RectangleInfo')
 
 
 class AffineMatrix(Structure):
@@ -178,6 +178,18 @@ CCMaxMetrics = 16
 
 
 class CCObjectInfo70A(Structure):
+    CCMaxMetrics = CCMaxMetrics
+    _fields_ = [('_id', c_ssize_t),
+                ('bounding_box', RectangleInfo),
+                ('color', PixelInfo),
+                ('centroid', PointInfo),
+                ('area', c_double),
+                ('census', c_double),
+                ('merge', c_bool),
+                ('metric', c_double * CCMaxMetrics)]
+
+
+class CCObjectInfo711(Structure):
     CCMaxMetrics = CCMaxMetrics
     _fields_ = [('_id', c_ssize_t),
                 ('bounding_box', RectangleInfo),